+++ src/modules/post/controller.ts
@@ -1,4 +1,5 @@
 import { Request, Response } from "express";
+import isUserOwn from "../../helpers/isUserOwn";
 import MessageHelper from "../../helpers/MessageHelper";
 import { PostService } from "./services";
 
@@ -48,6 +49,11 @@ export class PostController {
     createPost = async (req: Request, res: Response) => {
         try {
             const { title, content, image, authorId } = req.body as { title: string, content: string, image: string, authorId: number };
+
+            if (!isUserOwn(req.user.id, req.user.role.name, authorId)) {
+                return MessageHelper("You are not allowed to create post for this user", true, res);
+            }
+
             const post = await this.postService.createPost({
                 title, content, authorId, image
             });
@@ -62,9 +68,20 @@ export class PostController {
         try {
             const { id } = req.params as { id: string };
             const { title, content } = req.body as { title: string, content: string };
-            const post = await this.postService.updatePost(Number(id), { title, content });
+            
+            const post = await this.postService.getPostById(Number(id));
 
-            return MessageHelper(post, false, res);
+            if (!post) {
+                return MessageHelper("Post not found", true, res);
+            }
+
+            if (!isUserOwn(req.user.id, req.user.role.name, post.authorId)) {
+                return MessageHelper("You are not allowed to update this post", true, res);
+            }
+            
+            const postEdited = await this.postService.updatePost(Number(id), { title, content });
+
+            return MessageHelper(postEdited, false, res);
         } catch (error: any) {
             return MessageHelper(error.message, true, res);
         }
@@ -73,9 +90,20 @@ export class PostController {
     deletePost = async (req: Request, res: Response) => {
         try {
             const { id } = req.params as { id: string };
-            const post = await this.postService.deletePost(Number(id));
 
-            return MessageHelper(post, false, res);
+            const post = await this.postService.getPostById(Number(id));
+
+            if (!post) {
+                return MessageHelper("Post not found", true, res);
+            }
+
+            if (!isUserOwn(req.user.id, req.user.role.name, post.authorId)) {
+                return MessageHelper("You are not allowed to delete this post", true, res);
+            }
+
+            const postDeleted = await this.postService.deletePost(Number(id));
+
+            return MessageHelper(postDeleted, false, res);
         } catch (error: any) {
             return MessageHelper(error.message, true, res);
         }
